+++ HISTORY.rst
@@ -2,6 +2,13 @@
 History
 =======
 
+0.12.0 (2016-11-11)
+-------------------
+
+* Replace `Document.opts.children` by `offspring` and fix grand child
+  inheritance issue (see #66)
+* Fix dependancy since release of motor 1.0 with breaking API
+
 0.11.0 (2016-11-02)
 -------------------
 
@@ -47,6 +54,7 @@ History
 
 0.7.8 (2016-4-28)
 -----------------
+
 * Fix setup.py style preventing release of version 0.7.7
 
 0.7.7 (2016-4-28)